troupe: move non-primitive actions out of `Process`
Instead of keeping actions that can be built on top of primitive actions
in the `Troupe.Process` module, from where these higher-level actions
could touch implementation internals, move them to the `Troupe` module
instead.

See: https://github.com/NicolasT/troupe/issues/16
